requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.3.4
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.10.6
croniter==6.0.0
Django==5.2.5
django-cors-headers==4.7.0
django-redis==5.4.0
django-rq==3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
ffmpeg-python==0.2.0
future==1.0.0
gunicorn==21.2.0
iniconfig==2.1.0
kombu==5.5.4
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
redis==5.0.1
rq==2.5.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
